refactor: move offset_reader
